Changed the logic of LoomEnter(Exit)Thread
Inserting LoomExitThread after pthread_join doesn't work for Apache,
because one early-exit thread wouldn't be able to release LoomUpdateLock
until it gets reclaimed. Therefore, we still need to instrument thread
entries and exits.

We use IdentifyThreadFuncs to identify thread functions, and add
LoomEnterThread and LoomExitThread to their entries and exits. However,
IdentifyThreadFuncs is sound but not accurate in the sense that it may
mark some regular functions a thread function, and causes double
LoomEnterThread. Moreover, even if IdentifyThreadFuncs is 100% accurate,
the application may call a thread function as a regular function and
causes double LoomEnterThread as well.

We solve this issue by keeping track of CallDepth, a per-thread call
depth of marked thread functions. Whenever a function marked as a thread
function by IdentifyThreadFuncs gets called, we increment CallDepth;
when such a function exits, we decrement CallDepth. Note that CallDepth
is thread local, so we don't need any lock for it. We call
LoomEnterThread only when CallDepth changes from 0 to 1, and call
LoomExitThread only when CallDepth changes from 1 to 0.